requirements-dev.txt
django-comment-migrate

mypy
mypy-extensions
isort

flake8
mccabe
flake8-blind-except
flake8-broken-line
flake8-bugbear
flake8-builtins
flake8-class-attributes-order
flake8-cognitive-complexity
flake8-commas
flake8-comprehensions
flake8-debugger
flake8-eradicate
flake8-functions
flake8-isort
# flake8-mock>=0.3
flake8-mutable
flake8-print
flake8-quotes
flake8-string-format
flake8-variables-names
flake8-docstrings

autopep8

# https://black.readthedocs.io/en/stable/integrations/editors.html
black
types-requests
django-extensions
django-debug-toolbar
